Remove unused code.
Was trying to find a constant for Course model's active/inactive status. Found 
that GlobalConstantComponent has the constants, but  does not appear to be used
by anyone. Moved the STATUS_ACTIVE and STATUS_INACTIVE constants into the Course
model as proper class constants instead of vars and removed all references to 
GlobalConstantComponent.

Note that the User role constants should get the same treatment for converting
into class constants instead of vars. One complication is that the Role field
in the User table seems to be a bit redundant given the existence of the Role
tables. So don't want to touch that yet.

There were methods in the User model for assigning students and instructors
to a course, however, they were not being used. Assigning instructors can be 
done by the Course model's addInstructor() method. Enrolling students into 
courses was being done by queries. I've removed the code in the User model for
assigning instructors to a course. I've changed the Users controller to use the
enrolment methods instead of calling queries. Two new methods for adding and 
removing Roles from a User was added, they will be used in the LTI functions.

Also, might be a good idea to add foreign key constraints to the database 
creation SQL so we don't have to completely rely on CakePHP's constraint system.