Basic LTI, SSO and import working.
Courses can now be imported using LTI. Note that SSO currently only works with
users imported using LTI. This is due to SSO being dependent on a new field in
the Users table. The new field, 'lti_id' is used to store the id of the user
on the calling LTI application, since this id is probably different from the 
id iPeer uses.

The LTI url is http://your.site/lti

Removed the Faculty ACO stuff from the courses controller to keep the SVN copy
working. They were throwing an error about a table being missing.