Fixed and refactored installation wizard.
Note that the new installation wizard requires config/database.php to exist
before it can run.

Still TODO: the superadmin user is still being created with raw SQL queries.
This is NOT good practice for CakePHP and should be replaced with calling
CakePHP models. Unfortunately, it looks like there is a conflict in how we
store roles in iPeer. Roles are stored both in the Users table and the 
RolesUsers table. This needs to be fixed first.

Note, there might be a random error where the database configuration might
not be written to disk in time before the database patcher runs and uses
the pre-existing invalid credentials. Can't really seem to recreate it,
but have seen it twice, so might be a problem.

Hardcoded directories to config has been replaced with the CakePHP constant
CONFIGS, which should be more reliable.

The installation wizard now places a file called 'installed.txt' into the
CONFIGS directory to indicate that the installation was completed. This is
to fix a bug where you have to manually browse to /install to run the 
installation wizard. We now uses the router to direct users to the 
installation wizard if 'installed.txt' is not found.

The installation wizard should not exit php now if an error is encountered.

We now use the CakePHP 1.3 Session flash and redirect method to show errors
to users and to render the correct page in case of errors.

Files that were found to be using Windows line endings were converted to
Unix endings.