Fix database upgrade, installer bug.
The database upgrade wizard wasn't being triggered correctly due to
sysContainer incorrectly reading the system parameters from session.

The database upgrade wizard also wasn't completely migrated to 
cakephp 1.3, this was completed

The installer was incorrectly reading the database version variable 
as an array instead of as a scalar.

Added comment to the sysContainer component to phase it out. The way
it's implemented using Session variables introduces occasional edge
case bugs such as a freshly installed version of iPeer incorrectly
complaining that the database is out of date. This is fixed when you
relogin. It should be replaced by just directly referencing the 
SysParameter and SysFunctions models.