Db version now from SysParameter model.
Replaced all instances where sysContainer was used to get the
database version with using the SysParameter model directly.

sysContainer tries to cache the database version value in the
session. Unfortunately, while this works most of the time, it
can introduce bugs where the actual database version value
has increased, but the session cache still has the old value.

The InstallController now has its own layout called installer.
The default layout was bringing in a dependency on User that
wasn't necessary for the installer. Quite a bit more can
probably be removed from the installer layout, but that can
wait.