Convert table layout to css layout for install.
This is the start of the UI refactor to convert all the table layouts to
css layouts. Started with the installer since that's the one I'm most
familiar with. The plan is to refactor the installer first and then base
the rest of the refactoring on the installer.

The new installer layout is meant to be a starting point to modify the
default layout. I've moved the banner, debug, and footer sections into
their own elements. The new banner element is all CSS, no images except
for the ipeer logos. It gracefully degrades in IE8 to boxes.
Unfortunately, the position of the right block logo does not gracefully
degrade in IE6, but not going to worry about that.

The debug element has had the svn information removed. Note that when
refactoring the debug element, I noticed that the javascript library we
rely on, protoype, hasn't been updated for more than a year. Thus, I've
started migrating to jquery.

The new install.css file is meant as a template to replace the old
ipeer.css. The old ipeer.css file did not have a css reset section for
some reason?

Due to the banner being css styled now, I couldn't use the existing
banner images since they all had a built in transparent section or
border section built into the images. I've edited them to remove the
border sections and saved them with the _noborder postfix.