Installation step 2 now converted to css.
The GPL page was changed to not use the __('') strings since they were
breaking the layout by putting a horizontal scrollbar.

Each installation form for the next button appears to have different
heights, so the forms now have unique ids to set their heights in.

The heights are necessary since the floating Next button's height isn't
recognized by the browsers.