Convert install4 and install5 to css.
Install4 needed a model based form validation like install3, so the
install3 validator was renamed to install_validation_step3 and now we
have install_validation_step4.

The confirm password field in install4 has a custom validation function
that checks it is equal to the the password field. Have to check if the
password needs to be hashed.

Various fixes were necessary cause the form field names were changed.
E.g.: The data_setup_options weren't being picked up, since it was using
the old form names.

In the debug element, had to abandon the short form conditional
operators (:?), cause they didn't play well with echo. I wanted to pass
the output through the htmlspecialchars function so that the browser
displays the actual code in the debugging area instead of trying to
parse it, which required echo.