Convert default layout to css.
The installer layout and css was sort of a prototype for converting the
default layout. I've just basically rewrote the default layout based on
the installer layout.

The installer.css file now only has installer specific CSS. The
ipeer.css file now has all the site-wide css. There are large sections
of ipeer.css that still needs to be refactored, and I've marked them as
such.

There is now a navigation element that generates the navigation tabs.
Instead of incorporating the navigation into the banner, I've split it
out into it's own element so it's easier to deal wwith.