Refactored login and home page views.
Rename System Parameters and System Functions tabs to use the
abbreviation Sys instead of System. This frees up space in the
navigation bar. This is a fix for IE8 not having enough space to
display the username at the minimum resolution of 780px.

Note: The minimum resolution is 780px because we want to make sure
people with 800x480 phones or 800x600 monitors can still view the
site.

The W3C validator doesn't like tags such as 'onclick' to have
capitalized characters, e.g.: 'onClick' in the user_info.ctp case

Completely removed the rest of the old code from the default
layout. Also changed title to a header instead of a div.

iPeer Features Removed:
check whether user has email address when editing their own
profile (or submitting? not sure), ask them to enter an email if
they don't

Refactored the home page so that the code is easier to read.
Plus, fixed a bug where inactive courses weren't showing up.

Login page converted to css. With some generalized css rules for inputs
generated by the html helper.