Update popup layout and temp css fixes.
Just noticed that we have a popup layout. Updated that to be similar to
the default layout. Note that I removed the error message display
capabilities, will have to add those back if needed.

Also added temp css fixes so that the old refactored views still look
like they used to.