Remove tables, rewrote using CSS for course home.
The course home page, the page that lists a summary of the course and
the actions available, has been refactored to use CSS instead of tables.

Use the css classes course_actions, course_summary, and course_submenu
to control it.

Many of the course elements named ajaz_personalized_Course.SubMenu.*
were removed as they were abstracted out to the new submenu_items.ctp.

It was a bit redundant to have elements with file names like
course_submenu.ctp when they're already in the courses directory, so the
course_ prefix was removed.

The instructors.ctp file lists instructors as links to the instructor's
profile, so it was renamed "list_instructors.ctp" and formatted as a
html unordered list.

The home icon that used to be displayed next to course names was
removed, it seems a bit weird to hardcode the location of the home icon
into sys_container.php. If it should be added back, then it should be
done as a CSS background image.

Unfortunately, due to the imprecise nature of CSS' adaptable layouts,
there is a 1 pixel misalignment of the course action headers on the
right hand side where it doesn't match up with the previous boxes.