Popup layout CSS now looks like default layout.
While migrating the default layout to use CSS instead of tables, I gave
the popup layout a cursory conversion. This commit finishes the
conversion. The popup now look more like a mini version of the default
layout, with a rounded corners outer container and a smaller top rounded
banner.

The popup reuses the same css class as the default layout for the
outer container and banner.

It looks like pop-up sizes are hardcoded into links. I only changed one
of them, but might have to track down each of them to give a nice
looking size. Would be nice if the popups can get auto get a size that
fits.