CSS refactor, fix controller for users/add.
Change the form for users/add to use CSS instead of tables. This
involves the creation of the new USERS section in ipeer.css.

Adding the new user to courses should now work properly for at least
students and instructors. Not yet implemented for admins. Note that if
you want to preselect a course for the user to be added to, just pass in
the courseid of the course. E.g.: users/add/1 will preselect the course
with id 1.

Role selection for the new user is newly implemented. The old form was
writing to the deprecated roles column in the Users table. The new form
will properly update the RolesUsers table instead. Note that new users
created by this form will be broken since there are still places in the
code that checks for roles in the old style. We should replace the old
checks instead of writing in a placeholder role, easier to see the
presence of the outdated code if it's broken.

Send email notification calls the correct function, but not sure if the
email function actually works. Also not sure what email gets sent out,
need to test.

Form validation now requires a role, which defaults to student. The
previous form wasn't requiring those, which resulted in new users that
have no privileges and can't be deleted in the UI.

A first name is also now required, since some areas use the first and
last name as displays. If the user doesn't have a first or last name, it
doesn't fall back to the username, instead just displays a blank. This
behaviour proved unexpectedly complex to fix, so I just made this field
required instead.

Title and Student Number used to show up when the role is appropriate,
something that might be added back later, but isn't really critical at
the moment.

The old password generation was insecure, so I wrote a more secure one.
Unfortunately, this results in a password that was decidely less user
friendly. We can make it more user friendly by generating a shorter
password, if we can tolerate the weaker security. Still need to fix
password hashing.