Fix email notification when adding new user.
I setup a localhost email server to test the mail sending functionality,
unfortunately, the CakePHP validation rules rejected my localhost email
address. So I had to change the email validation to a custom one. If you
have PHP >= 5.2.0, it'll use PHP's filter_var() to validate emails. If
you don't, it'll use a really barebones regex to make sure you have text
on both sides of the @ sign.

During new user creation, there is an option to email the user being
created about their account credentials. Interestingly enough, there
doesn't seem to be a way to configure the email user except directly
modifying the database, so this will have to be fixed later as an
installation step.

If the email server is configured correctly, messages do get sent out.
The addUser mail template needed to be rewritten to provide more
information to the recipient. The recipient will now know their
username, password, and courses they have been enrolled in, along with a
link to the site.