Fix navigation generation not using new roles.
The navigation bar was generating its menus based on the old role column
in the user table. This was changed to be based on the new RolesUsers
table.

Unfortunately, since we don't have the new access control system fleshed
out yet, the permissions were just hardcoded. So this will have to be
fixed when the new permissions system is done.