Fix bad use of read() in user model.
The read() function apparently doesn't just return one entry as I
thought, it returns that entry but also sets the the current user as the
model's current record. This results in things like creating a new user
as root would override root's information with the new user's
information. The findBy() function family was what I wanted.