Fix for add student not adding students to course.
Previously, when you click on Add Student in courses/home and follow the
form, the student created is not added to the course in question. This
was partially due to a broken add user form and needed the add user form
to be fixed first and was responsible for most of the previous commits.
Now that the add user form properly adds users and instructors to
courses, we can finally fix the original bug.

Now when you click Add Student, the course that originated the request
will be automatically selected as a course to enrol the user in.