Fix not being able to list users by course.
This is based around goToClassList in users controller. The old code
just adds a class filter and then redirects to index.  The index uses
the old prototype based ajaxList, I decided to switch to jQuery based
DataTable. The switch was done mostly because the ajaxList code to
filter based on class wasn't working and a long look at the ajaxList
code base produced only puzzlement. As we want to move off of prototype
and to jquery anyways, it was easier to just rewrite this part.

The only problem with the DataTable migration is speed. The way
DataTable works is to convert an existing table with all the data into a
paginated table. If there are a lot of users, then this might be a bit
slow.  E.g.: current production has about 10 000 users.  But we might
expect some 60 000 users later. While the loading speed was fine for 10
000, I have doubts about handling 60 000 users. The ajaxList code
retrieves users on demand, so does not have this problem.

So the index was kept on the old ajaxList code.