Remove database.php from version control.
Code was added to bootstrap.php that creates an empty database.php file
if it couldn't find one.

database.php was initially added to version control since CakePHP1.3
errors out if it can't find it. But it's proven to be a pain, since each
dev instance can have different credentials and switching between
branches overrides the existing database.php.

This fix lets us have the best of both worlds.