goToClassList user menu w/ view, edit, email, etc.
A right click on a user in the users index page would present a menu
with view, edit, email, reset password, and delete options.
Unfortunately, this right click menu isn't automated test friendly. The
new classlist table use an alternate method where it inserts a new row
into the table with the controls instead.

The addition of the controls requires the user id to be added to the
data passed to the view.

When an action is done, users were hardcoded to be redirected to the
index. This was changed to redirect to the referer, so if actions were
initiated from the classlist, they would be redirected back to the
classlist instead.

Two bugs were spotted when working on this:

1. Reset password needs to be switched to the secure password generation
class.

2. View user for some reason uses the add user layout.

These still needs to be fixed.