fixed the IE not resizable popup problem
Removed the section of the code in the if else structure for newWindow -
including the window.open portion. This avoids the problem of it being
used instead of wopen in ipeer.js. What's left in the if else structure
in ajaxList.js is window.location if it's NOT opening a new window.
In the ipeer.js file, I removed the commentd out code in wopen() to
clean up the code.