Fix mass email wizard.
The mass email wizard wasn't working due to conflicting ways of how it
tries to obtain the list of target email addresses. Namely, it expects
different kinds of params to emailer/write/ based on which type of email
you're sending.

For users, it assume there will only be 1 param, the user id.  For
courses, it assumes there will be 2 param, the class id and the
character 'C'.  For groups, it concatenates the 2 params into 1, and you
get a param like 'G1'.

This was fixed by making everyone consistent, the emailer/write page
will accept 2 params, the type and id.

Type: U - User C - Course G - Group

Other than that, the email is properly saved to the scheduler.
Unfortunately, it looks like there is no system to properly send out the
email at the scheduled time. You can access the emailer/send page to
manually trigger a send.

A very basic way to get scheduled mail working would be a cron job that
simply wgets the emailer/send page. We can also rig up a scheduled mail
check every time someone visits the site.