Allow unauthed access to emailer/send.
Emailer/send is used to check if there are any pending messages to sent
out. In order to enabled scheduled message sending, there needs to be a
cron job that accesses that page periodically. I used something like
this in crontab:

* * * * * /usr/bin/wget http://localhost:2001/emailer/send

wget won't work if we need to login before accessing that page.