Fix add mixeval error and success messages.
The error message wasn't showing up cause it was trying to use a
non-existing error message property of the mixeval controller. Looks
like the property was changed to a method call.

The success message was changed to use the 'good' message layout, with
the green theme instead of the red.