Fix course listing for searchEvalution.
The course list wasn't being set properly. Unfortunately, it looks like
the search itself doesn't work. On Michael's computer, it at least
returns all events, but doesn't do any other filtering. On my computer,
it just seems to do nothing.

This probably means that search need to be rewritten. But since it's a
lower priority, going to skip that for now.