fixed the bug of not making full copy of Mixeval
Previously when we copy the Mixeval the general information (eg. name)
is not autofilled. However for the questions most of the data for the
questions are autofilled correctly, except for the Lickert questions'
descriptors.

I've found that the first problem was that the original read() function
was not able to pickup the descriptors. I fixed that by using the find
function instead. I have also fixed the autofilled name to be
"Copy for ... " instead of blank.