fixed the bug - can't view Creator in Mixeval
This bug is the same as the one John fixed earlier for rubrics
where we can't view the creator information when we click on the
creator's name or "View Creator".

We change it to read from the right table - " from table Creator
and column id"  to "from table Mixeval and column creator_id".