IE8 compat for evalviewer, css fix for mixeval.
Evalviewer suffers from the same problem in IE8 as evaleditor where the
keyword 'class' can't be used bare as a key in a hash. It just needs to
be changed to a string.

Also removed some javascript hardcoded styles in evalviewer for the
scale weight.

The scale weight field was being wrapped and positioned too low in a way
that it overflows the question box boundary. This was fixed so that it's
positioned in the same line as the 'Descriptor' text.