No redirect in add email tmpl, flash msg in popup.
When saving using the Add Email Template popup, it redirects the popup
to the index page if successful. This was removed to just display a
flash message instead.

It was found that flash messages were included in the popup layout, so
that was added. The popup layout was also changed to use 4 spaces tabs
instead of 2 spaces. Sections are now delineated by comments.