changed the edit survey to display course code
Previously the edit page for surveys had a drop down menu that
allows the user to change the course for the survey, but it was
displayed as the course's title instead of the course code which
more people will recognize.

I've now changed the drop down menu to display the course code
instead.