fixed the broken add/edit courses
Previously, the add and edit functions for courses were broken.
To fix it, I redid the forms using cakephp instead of the original
javascript. By doing so, we could use the validation and remove the
duplicateName and beforeSave functions.

I have also edited the ipeer.css file to make the new forms look
better. The instructors element in the edit function is now a
selectbox. So the user will have to hold on to command or control to
select multiple instructors.

I have also added the validation to make sure the Course entered is
unique and not empty.