fixed redirecting homepage for students
There was a small bug where students are not redirected to the
Evaluations page even after they have entered their email. They are
only redirected to the Evaluations page if they logout and log back
in.

I found that even though the database is updated instantly, the function
$this->Auth->user('email') is not picking up the new data. Therefore I
fixed the bug by grabbing the data from User array.