fixed - students can't see their evaluations
When we log in as a student, we could not see the evaluations because
the variable data could not be passed to studentIndex which is being
rendered from index. I've added a function studentIndex to specifically set the data variable.

For some reasons the UserEnrol model does not seem to work really well, which
was being used for retrieving course_ids for the enroled courses. Instead, I've
changed it to use the User model instead.

Lastly, studentIndex was trying to use a function called formatDate, which
doesn't seem to exist anymore. Instead, I've changed it to use date(), a php
function. For better visual, I've also changed the table slightly to view
the date in one line instead of two.