added new permissions to fix eval toolbar
When we started using the new permissions mechanism, it broke the
evaluation toolbar.

I've added new permission for each of the different controllers for
the evaluations so that they can be more flexible for new roles we may
be implementing.

I've also cleaned the code for tools_menu.ctp by creating a function,
generateLink, to create each link for the toolbar instead of repeating
the code for each link. This is very similar to what we did for the tabs.