Add new models for faculty admin.
This is to store the faculty admin role. Works like this:

Faculties can have many departments. Each department can have many
courses. Each department can belong to only 1 faculty. But each course
can belong to many different departments.

A user can belong to many faculties.

Current plan:

When requesting a page that requires admin access, 2 levels of check are
performed:

1. To see if you have admin access.
2. To see if you have access to the department.

Access to the department is granted if you are in the same faculty as
the department.