fixed the filter for user's evaluation tools
There was a bug where the filter for showing only the evaluation tools
the user has created would not work. Instead a pop-up appears (Firefox)
saying that it may be due to network down or invalid server response.

I've fixed the problem for emailtemplates, mixevals, rubrics, and
simpleevaluations by changing the 'id' for $jointTableCreator from
Creator.id to Creator_id. The reason is the the error that also appeared
said that Creator.id is an unknown column.