fixed missing add evaltools buttons
In the "All My Tools" area, only Simple Evaluations has an add option.
While Rubrics and Mixed Evaluations didn't. I found that the two buttons
that were misssing was because they were enclosed by an if statement using
the access function (for permissions), which is not used anymore.

All I did was removed the if statement, and the buttons/links returned.