fixed the issue with copying mixed evaluations
Previously, when copying mixed evaluations, the data for the original
mixed evaluation would be "erased" when the new copy is made.

To fix this issue, I did what was done for copying rubrics and simple
evaluations. I cleaned up the array by removing any trace of ids, users'
names and created/modified times.

I've also changed the successful messages to be green.