Admin tab in navigation.
With the addition of the Faculty and Department pages, it was getting a
bit crowded in the navigation. The Sys Parameters and Sys Functions tab
were taken down and replaced with the Admin tab.

The Admin page is a static page with a list of available admin
functions: Faculties, Departments, Sys Params, Sys Functions. Only the
superadmin will see all 4 options, while the faculty admin would see
only the Department link.

Note that in order to add the admin page to ACL, I had to break
convention since 'pages/admin' wasn't working. It was conflicting with a
pages controller, probably brought in from the PHP core. The permission
alias for the admin page is instead just 'adminpage'.

The sql template files were updated with the appropriate ACL entries and
department/faculty samples.