Add/Edit user now has a Faculty field for admins.
Part of the Faculty Admin implementation. Users can now be created and
edited to have a faculty. Note that the faculty field is being soft
limited to admins currently. Other users may have a faculty entry, but
nothing will happen, since the faculty entries are currently only used
for limiting access to admins.

The faculty field only shows up for users with access to
'functions/user/admin', which is currently only the superadmin. The
faculty field is disabled unless 'admin' is selected for the role.