Alter the permissions slightly for users
I have changed the permissions to allow users to be able to view users
with the same role but not allow them to add, edit, or delete users that
have the same role.

To do this, I editted the permissions to go down one more level with create,
read, and delete and added a getViewableRoles() function in access_control.php.
I've added getViewableRoles to allow filtering for users with the same role
as the user in the User index but not allow them to add users with the same
role.