Department selection for courses.
At least one department is now required to be selected for courses. If
you are a superadmin, all departments will be available for selection.
If you are an admin, only the departments in your faculty will be
available for selection.

Also added departments to the sample data courses.

Also removed some unnecessary instructor selection code since the
CakePHP magic stuff seemed sufficient.