Fix failing course test cases.
The Course->addInstructor() function used to test the logged in user's
role to see if they are permitted to add instructors. However, since it
uses the old role column in the Users table, it doesn't work now.
Unfortunately, it seems that User::hasPermission doesn't work in test
cases, probably cause it needs the ACL tables. So this check and the
corresponding test case was removed. It should be fine, since the
controller that calls the model methods should check permissions
properly already.

The validCourseInstructor tests in testDeleteInstructor() was removed
since they didn't seem needed to test delete instructor specifically.