fixed the survey result interface
Previously, nothing was showing up correctly for the survey result page.
The problem seemed to be with the pagination. I've tried to fix it with
cakephp Paginator and ajaxList but couldn't get either of them to work.

The cakephp Paginator has problems with remembering the data for page 2
and onwards after filtering. I believe I couldn't get the filtering to
work properly for ajaxlist because another ajaxlist already exists for
the survey_group controller.

In the end, I have used the dataTable which is the same one being used for
the class list.

There are no actions (eg. viewing individual results) for the interface
 (except for search) yet. I will be implementing them next