fixed the filtering for evaluations index
Previously the filtering for Self Evals and Comments were not "fully"
working. They were able filter "Enabled" and "Required" respectively.
However, the filtering for "Disabled" and "Optional" returns "All".

Therefore I increased the value for the options by 1 (eg. from 0 to 1
and 1 to 2). This slight change fixed the filtering. It seems like the
ajax list does not like the value 0.

To ensure no bugs appear, I've also changed the events add view to assign
1 and 2 to the options instead of 0 and 1.