fixed a bug from the previous commit
Firstly, I've commited the evaluation controller instead of the events
controller, which I have added to this commit.

Secondly, I've fixed the bug when we add surveys. The default value added
for comments and self_evals were set to 0, which isn't used anymore. It
has not been corrected to use 1.