changed the link on the evaluations page to go to events index
Previously, the link goes to the evaluation index, which doesn't work.
So instead I have now used evaluations index to redirect to events index.

I have also added an extra (optional) parameter, which is the course id of
the current course or current event. Three different situations can happen
with this parameter. Firstly, if a valid id is inputted, the events index
will filter and display only the events for that course. Secondly, if no
course id is inputted, then the events index will display all the events in
the courses the user (not student) has access to. Lastly, if a course id that
is not numeric or not a valid id, "No Result" will be displayed.

I have also commented out the goToClassList function in the events
controller, because I  have placed that bit of code in the index itself.