added permission restrictions to evaltools and simple evaluations
For evaltools, as long as the user has permission for the controller
evaltools, they can access their list of evaluation tools.

For simple evaluations, the user must have the  minimum requirement of
having permission for the controller simpleevaluation. The id entered must
also be numeric and valid (except for add and index).

For edit and delete, the user must also be at least one of the following:
creator, admin, or superadmin. For edit, there must also be no sumissions
in any of the events using the simple evaluation. For delete, no events can
exist using that particular simple evaluation.