added permission restrictions for emailer
index - simply is a check if the user has access to the controller - emailer

write - if the $id entered is invalid (for course, group, or user), the
user will be redirected back to the index. I've also used the new permissions
functions/email/all... for checking if the user has permission to email
the specific class, group, or user with $id.

cancel - checked that the $id is valid and the user has permission to cancel
the schedule - creator, admin, or superadmin

view - same as delete

I've also added the 'good' attribute to the successful messages.