fixed the students list for adding/editing groups + redirect for survey
Previously, the list of the students (in selected group and rest of class)
are not displaying in the form for creating/editing group.

For the students that are already assigned to other groups that needs an
asterix appended to their name - I've "find" them separately from the rest
of the class. Then merge the two lists together.

find('list') is the function we want to use (not find('all')), because it
helps making the edit work properly. If we use find('all') and then filter
it - the edit won't work properly - eg. if we just edit the name, the group
would be changed to having users with id 0 to 2 as members (eg. for 3 students)
In other words the array index was used as the id instead of the actual id.

find('list') wasn't working properly before - it ignores the "Group" and
"Enrolment" arrays we've "contained". To fix this, I've added the parameter
"recursive => 1".

Minor change - after copying a survey, I've redirected it to the index.