fixing the unstable build
I have changed the implementation of getStudentsNotinGroup in group model.
Therefore I had to change the test cases(where 2 of the errors occur). The
test cases now change student_no_with_full_name & asterix.

Previously, if the user has no student no, first name or last name,
student_no_with_full_name would be null (empty). So we have now used
a virtualFields variable that uses CONCAT_WS for full_name and
student_no_with_full_name instead.

One of the erros in the build was that the due date test. It so happened that
the due date was set to June 10, 2012, which just past. So we have now changed
it to June 10, 2022.