removed the message from course index
The message was about the result if the user searches by instructor.
However, we do not have the option to search by instructor any longer.

Right now the result total reflects the total of the instructors of the
courses in the result instead of the total courses in the result. I will
fix this soon.