fixed events index, penalty from events add and edit
Some of the redirect links after adding event for example has a message
appeneded to the url, which caused the index not to work, which I fixed.
I've also fixed some links that were not working.

I've also fixed penalty, that seemed to be set to active whether we choose
to make it active or not when we add an event. It now only adds penalty to
the penalties table when "Yes" is selected. Also when a user edits their
event from having penalty to NOT have penalty, the entries in penalties
will be deleted.

I've also added a field in the view events page to view the penalty details.