used 0 and 1 again for events self_eval and com_req
Fixed the problem in ajax_list.php where it is checking whether the $value
(eg. 0 or 1) is empty or not. It so happens that 0 is considered empty. I've
fixed this by checking $value to not equal null and not equal "".

I've also undo all the changes I made previously by changing the values from
1&2 back to 0&1 respectively.

The other changes I made in this commit is for some visual problems.