removed model UserGradePenalty and penalty is now calculated at real time
When I tried to implement penalty for students that don't submit evaluation,
I find that it was very hard to implement a mechanism that can add the penalty
to the database. Therefore it was decided to calculate all the penalties on the
fly.

I've have removed the UserGradePenalty fixture and model. I've also changed  all
code that depends on the model UserGradePenalty to code that will calculate the
penalty on the fly.

The "formatPenaltyArray" has been moved to Evaluation_Base model instead so that
all the evaluations can access it. The code that calculates the penalty on the fly
is similar to the code I've used for "formatPenaltyArray".

Along the way, I've also fixed invalid functions/variable bugs for exporting to
csv and excel. These were very simple fixes.

I've been using simple evaluation to test the penalties, therefore I would be doing
further tests using mixeval and rubrics to fix any bugs that might exist.