finished implementing penalty to the mixeval result
I've added a scrollbar to the summary table in the detailed view, because
as more questions are added, there would not be enough room. To make the
scrollbar work, I had to add some new classes to ipeer.css to add new
attributes and to override some attributes. Scrollbar only appears when
needed. Only the horizontal scrollbar is implemented. The vertical
scrollbar has been disabled.

I've also editted the Final Average for the individual student to average
percentage received per question. the average is calculated based on ratio.
This means a question with a weight of 5 will be worth more in the average
than a question that has a weight of 1.