implemented penalty for the rubric results
Like the mixeval results page, I've implemented penalty to the rubric
result page. I've also added a scrollable section on the top of the
detail result page to allow more room for more questions.

I've noticed that the <strong> tag wasn't working - meaning all the
"titles" were the same as the rest of the text. I've added
font-weight: bold for strong in ipeer.css. If it is too bold, we can try
adjust it a bit.