fixed student view simple eval result
I've fixed comments release, which display all comments as long as the
first comment in the array is released. It will not check release status
for each comment before releasing it to the students.

Fixed the student index for displaying submitted evaluations. Before, they
will display once the evaluation release date is past. Now they will be there
until result release date is past.

I've also fixed releasing individual comments on the instructor side. It was
a simple Query syntax error.

Lastly, for the student view of simple eval result, I've changed the note
from just saying the red star means penalty. I've now changed it to only be
shown when a penalty is given. It shows the percentage of the penalty.