implemented penalty to student view of mixeval result
I've added penalty to the rating for mixeval. The rating is also something
new I've added in this commit. The result now not only show the Points but
also the Grade.

To make things fit on the screen, I have changed the results to be on the
regular screen instead of pop-up.