implemented penalties to the csv and excel export of evaluations
The penalties are now implementd. However there are many minor issues
that need to fixed such as comments not placed with the correct person.

Also the implementation of export to excel and csv (excel expecially) seem
to be more complicated than necessary. So if we have time we may want to
reformat the export and redo the code.